// apps/web/src/lib/api/generated/main/model/loot-stats-response-dto-output-top-npcs-item.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { LootStatsResponseDtoOutputTopNpcsItemByRarity } from './loot-stats-response-dto-output-top-npcs-item-by-rarity';
import type { LootStatsResponseDtoOutputTopNpcsItemType } from './loot-stats-response-dto-output-top-npcs-item-type';

export type LootStatsResponseDtoOutputTopNpcsItem = {
  npcId: number;
  name: string;
  /** @nullable */
  type: LootStatsResponseDtoOutputTopNpcsItemType;
  /** @nullable */
  lvl: number | null;
  /** @nullable */
  icon: string | null;
  count: number;
  byRarity: LootStatsResponseDtoOutputTopNpcsItemByRarity;
};